Add files via upload
ESPN file is actually quite useful and the bulk of the upload. Can scrape specific games, years or whole seasons.
Kenpom does a yearly and daily scrape. Should only be done once so you have the daily info and can be periodically updated.
Other two files are specific to the 2022 march madness tournament.